Bug fixes and some additions
This commit will solve the problem of
repeating questions and ensure that
the number of questions input by the
user is less than or equal to the number of
possible combinations of a range of numbers
inclusive of maximum number and
minimum number.

There is a label text showing to enter no. of questions below 56
if the user chooses
very easy as the difficulty level, else it will not be shown.